logout in navbar and dashboard link

+++ nextjs-sunny-pg/src/app/user-dashboard/helpers/helper.tsx
@@ -40,7 +40,10 @@ export const uploadPhoto = async (
 ) => {
   const fileName = `${userId}-${type} -${file.name}`;
   const filePath = `${pathName}${fileName}`;
-
+  if (file.size > 150 * 1024) {
+    alert("File size is large! Please select photo with smaller file size");
+    return null;
+  }
   try {
     await deleteAllFiles("photo-ids", pathName);
